requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain-deepseek>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.25.0
python-multipart>=0.0.6
requests>=2.31.0 